% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinyin.R
\name{pinyin}
\alias{pinyin}
\title{Obtain and label Chinese pinyin}
\usage{
pinyin(input)
}
\arguments{
\item{input}{A string consisting of Chinese character or sentences.}
}
\value{
A string consisting of pinyin and input Chinese character.
}
\description{
When you pass in a Chinese character, you can obtain the pinyin of the
Chinese character, so that you can more easily understand the pronunciation
of the Chinese character.
}
\examples{
# Example(Not run)
# pinyin("type Chinese character")

}
