% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_view.R
\name{translate_view}
\alias{translate_view}
\title{Open a Youdao website browse translation results}
\usage{
translate_view(input, from = "auto", to = "auto")
}
\arguments{
\item{input}{An English or Chinese word.}

\item{from}{The source language, an optional parameter.}

\item{to}{The target language, an optional parameter.}
}
\value{
A list consisting of Phonetic, explains, etc about target language.
}
\description{
When you pass in an English or Chinese word, this function will Open
Youdao website browse translation results.
}
\examples{

# Example(Not run)
# translate_view("good")
# translate_view("quarto", from = "en", to = "zh-CHS")

}
