% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate English sentence into Chinese, or translate Chinese sentence into English}
\usage{
translate(input, from = "auto", to = "auto")
}
\arguments{
\item{input}{An English or Chinese sentence.}

\item{from}{The source language, an optional parameter.}

\item{to}{The target language, an optional parameter.}
}
\value{
The translation results about target language.
}
\description{
When you pass in an English or Chinese sentence, this function will calls
the Youdao text translation API for R to return the corresponding type
of Chinese or English representation.
}
\examples{
\dontrun{
  translate("I like China")
  translate("quarto", from = "en", to = "zh-CHS")
}

}
