% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate English words into Chinese, or translate Chinese words into English}
\usage{
translate(input, from = "auto", to = "auto")
}
\arguments{
\item{input}{An English or Chinese word.}

\item{from}{The source language, an optional parameter.}

\item{to}{The target language, an optional parameter.}
}
\value{
A list consisting of Phonetic, explains, etc about target language.
}
\description{
When you pass in an English or Chinese word, this function will calls
the Youdao text translation API for R to return the corresponding type
of Chinese or English representation.
}
\examples{
# Example(Not run)
# translate("good")
# translate("quarto", from = "en", to = "zh-CHS")

}
