% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_view.R
\name{translate_view}
\alias{translate_view}
\title{Open a Youdao website browse translation results}
\usage{
translate_view(x)
}
\arguments{
\item{x}{A vector made up of English or Chinese words}
}
\value{
Not print any result but open a website
}
\description{
When you pass in a vector consisting of an English or Chinese words,
this function will Open Youdao website browse translation results.
}
\examples{

# Example-1
translate_view("apple")

# Example-2
x <- c("apple", "banana", "pear", "I love you")
translate_view(x)

}
